refactor(frontend): restructure project and improve state & data handling
