+++ frontend/src/components/CartSidebar.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import { useCart } from "./CartContext";
+import { useCart } from "../context/CartContext";
 
 type CartItem = {
   id: string;
@@ -30,7 +30,6 @@ export default function CartSidebar() {
 
   useEffect(fetchCart, [reloadFlag]);
 
-  /* 修改数量、删除、清空 */
   const updateQuantity = async (id: string, q: number) => {
     if (q < 1) return;
     await fetch(`/api/cart/${id}/quantity?quantity=${q}`, { method: "PUT" });
@@ -48,7 +47,6 @@ export default function CartSidebar() {
     fetchCart();
   };
 
-  /* 按餐厅名分组 */
   const grouped = cart.reduce((acc, c) => {
     const key = c.restaurantId;
     (acc[key] ||= { name: c.restaurantName ?? `餐厅 ${key}`, items: [] });
@@ -64,7 +62,6 @@ export default function CartSidebar() {
       padding: "1rem", height: "100vh", overflowY: "auto", position: "sticky", top: 0
     }}>
       <h3>🛒 我的购物车</h3>
-
       {loading ? <p>加载中…</p> :
         cart.length === 0 ? <p>暂无商品</p> :
           Object.values(grouped).map(g => (
@@ -86,12 +83,9 @@ export default function CartSidebar() {
                 ))}
               </ul>
             </div>
-          ))
-      }
-
+          ))}
       <hr />
       <p><strong>总计：</strong> €{total.toFixed(2)}</p>
-
       <div style={{ display: "flex", justifyContent: "space-between" }}>
         <button onClick={clearCart}>清空</button>
         <button style={{ background: "#007bff", color: "#fff" }}>结算</button>
