+++ frontend/src/context/UserContext.tsx
@@ -1,0 +1,39 @@
+import { createContext, useContext, useEffect, useState, ReactNode } from 'react';
+
+type UserContextType = {
+  username: string | null;
+  setUsername: (name: string | null) => void;
+};
+
+const UserContext = createContext<UserContextType | undefined>(undefined);
+
+export function UserProvider({ children }: { children: ReactNode }) {
+  const [username, setUsername] = useState<string | null>(null);
+
+  useEffect(() => {
+    const token = localStorage.getItem("token");
+    if (token) {
+      fetch("/api/users/me", {
+        headers: { Authorization: `Bearer ${token}` },
+      })
+        .then(res => res.ok ? res.json() : Promise.reject())
+        .then(data => setUsername(data.username))
+        .catch(() => {
+          localStorage.removeItem("token");
+          setUsername(null);
+        });
+    }
+  }, []);
+
+  return (
+    <UserContext.Provider value={{ username, setUsername }}>
+      {children}
+    </UserContext.Provider>
+  );
+}
+
+export function useUser() {
+  const context = useContext(UserContext);
+  if (!context) throw new Error("useUser must be used within a UserProvider");
+  return context;
+}
