+++ frontend/src/layouts/MainLayout.tsx
@@ -1,12 +1,11 @@
-// src/MainLayout.tsx
 import { Outlet } from "react-router-dom";
-import CartSidebar from "./CartSidebar";
+import CartSidebar from "../components/CartSidebar";
 
 export default function MainLayout() {
   return (
     <div style={{ display: "flex" }}>
       <div style={{ flex: 1 }}>
-        <Outlet /> {/* 渲染当前子页面（App or MenuPage） */}
+        <Outlet />
       </div>
       <CartSidebar />
     </div>
